Update BaseNumber.java
Wrote two new instance methods that checks if the base is in a base given by the user called is(Base).  And the other method isValidBase(Base) that checks if the number is a valid in its given base and checks if the given base is equal to the given base from the user.